Move expression handling into its own module
